# internal/web/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: harbrr management API
  version: 0.1.0
  description: >
    harbrr's HTTP surface: the management API (first-run setup, authentication,
    API-key management, indexer and app-sync configuration) plus the entry points of
    the *arr-facing Torznab feed. The Torznab wire contract itself — the XML body
    shapes — is defined by Jackett/Newznab and served by internal/torznab; here we
    document only its HTTP routes, parameters, and content types (see
    docs/architecture.md). Hand-authored, embedded via //go:embed, and guarded by a
    drift test that asserts the spec matches the mounted routes (make test-openapi).

    Authentication: the management API uses a cookie session (after login) or the
    X-API-Key header; the Torznab feed authenticates with an apikey (or passkey)
    query parameter, as *arr clients expect. CSRF: a cookie-authenticated mutating
    request (POST/PUT/PATCH/DELETE) must echo the session's CSRF token in an
    X-CSRF-Token header — the token is minted at login and handed to the client via
    the non-HttpOnly harbrr_csrf cookie and GET /api/auth/me; a missing or wrong
    token is 403. X-API-Key and auth-disabled callers are exempt (see
    docs/security.md).
  license:
    name: GPL-2.0-or-later
    url: https://www.gnu.org/licenses/gpl-2.0.html
...
paths:
...
  /api/app-connections:
    get:
      summary: List *arr/qui app-sync connections
      operationId: listConnections
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      responses:
        "200":
          description: the configured connections (app key redacted)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AppConnection" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Add an app-sync connection (mints a dedicated harbrr key)
      operationId: createConnection
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateConnection" }
      responses:
        "201":
          description: the created connection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppConnection" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/app-connections/sync:
    post:
      summary: Reconcile every app-sync connection now
      operationId: syncAllConnections
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      responses:
        "200":
          description: per-connection sync outcomes (one entry per connection)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ConnectionSyncResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/app-connections/{id}:
    get:
      summary: Get an app-sync connection (app key redacted)
      operationId: getConnection
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: the connection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppConnection" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: Update an app-sync connection (a new apiKey rotates the app credential)
      operationId: updateConnection
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateConnection" }
      responses:
        "204": { description: updated }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      summary: Delete an app-sync connection (revokes its minted key)
      operationId: deleteConnection
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "204": { description: deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/app-connections/{id}/enable:
    post:
      summary: Enable an app-sync connection
      operationId: enableConnection
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "204": { description: enabled }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/app-connections/{id}/disable:
    post:
      summary: Disable an app-sync connection
      operationId: disableConnection
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "204": { description: disabled }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/app-connections/{id}/test:
    post:
      summary: Test a connection's reachability and credentials
      operationId: testConnection
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: the test result (ok=false carries a scrubbed error)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TestResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/app-connections/{id}/sync:
    post:
      summary: Reconcile harbrr's indexers into the connection's app now
      operationId: syncConnection
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: the per-indexer sync outcomes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncReport" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/app-connections/{id}/status:
    get:
      summary: Get a connection plus its per-indexer sync ledger
      operationId: connectionStatus
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: the connection with its ledger
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConnectionStatus" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/app-connections/{id}/indexers:
    put:
      summary: Set the selected-indexer set (for index_scope=selected)
      description: >
        Replaces the connection's selected indexers: the given instance ids become
        selected, every other currently-selected one is cleared. Only meaningful when
        the connection's index_scope is "selected".
      operationId: setConnectionIndexers
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SelectIndexers" }
      responses:
        "204": { description: selection updated }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: harbrr_session
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        A bearer credential — send over HTTPS only; it must never traverse a
        cleartext link. Terminate TLS at harbrr or a trusted reverse proxy.
  responses:
    BadRequest:
      description: invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: authentication required or failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: conflicting state
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: human-readable message }
        code:
          type: string
          description: >
            machine-readable code clients can branch on — one of: bad_request,
            invalid, unauthorized, invalid_credentials, invalid_api_key, forbidden,
            not_found, conflict, already_setup, not_implemented, internal. An OIDC
            ID-token verification failure also uses invalid_credentials.
      required: [error, code]
...
    TestResult:
      type: object
      description: >
        The outcome of a test action. ok=false carries a credential-scrubbed error
        message; the HTTP status is 200 even on a credential/connectivity failure.
      properties:
        ok: { type: boolean }
        error: { type: string }
      required: [ok]
...
    AppConnection:
      type: object
      description: >
        An *arr/qui app harbrr syncs its indexers into. The app's API key (apiKey) is
        never echoed — it reads back as the <redacted> sentinel.
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        kind: { type: string, enum: [sonarr, radarr, lidarr, readarr, whisparr, qui] }
        appId: { type: integer, format: int64, nullable: true, description: the App holding this connection's identity + credential (ADR 0004) }
        baseUrl: { type: string, description: the app's own API base URL (enriched from the App) }
        harbrrUrl: { type: string, description: the base URL this app uses to reach harbrr's feed (from the App) }
        apiKey: { type: string, description: "the app's API key, redacted (<redacted>)" }
        enabled: { type: boolean }
        syncLevel:
          type: string
          enum: [full, add_update]
          description: "full = add+update+remove orphans; add_update = add+update, never delete"
        indexScope:
          type: string
          enum: [all, selected]
          description: which harbrr indexers this connection mirrors
        freeleechMode:
          type: string
          enum: [honor, bypass]
          description: >
            which feed variant is pushed: honor = the standard feed (the indexer's
            freeleech setting is respected); bypass = the /full variant (full catalog,
            for cross-seed). Defaults by kind (qui = bypass; *arrs = honor).
        priority: { type: integer }
        syncProfileId:
          type: integer
          format: int64
          nullable: true
          description: >
            the sync profile this connection uses, or null for none (default behavior).
            ON DELETE SET NULL — deleting the profile drops the reference. Never set for
            qui (sync profiles do not apply to it); a profile with a non-empty category
            set must overlap the app's content range (the attach-time overlap guard).
        lastSyncAt: { type: string, format: date-time }
        lastSyncStatus: { type: string, description: "ok | partial | error | skipped" }
        lastSyncError: { type: string, description: scrubbed summary of the last failure }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, kind, baseUrl, harbrrUrl, enabled, syncLevel, indexScope, freeleechMode, priority, createdAt, updatedAt]
    CreateConnection:
      type: object
      description: >
        References the App holding the app's identity + credential either by appId
        (reuse an existing App) OR inline (baseUrl + apiKey [+ username], which
        get-or-creates the App by (kind, base_url)). harbrrUrl is required on the inline
        path (it backfills the App's feed URL).
      properties:
        name: { type: string }
        kind: { type: string, enum: [sonarr, radarr, lidarr, readarr, whisparr, qui] }
        appId: { type: integer, format: int64, description: reuse an existing App (omit baseUrl/apiKey/harbrrUrl) }
        baseUrl: { type: string, description: inline App identity (get-or-create) }
        apiKey: { type: string, description: the app's API key (sealed on the App) }
        username: { type: string, description: for user+password apps (empty for API-key apps) }
        harbrrUrl: { type: string }
        syncLevel: { type: string, enum: [full, add_update], default: full }
        indexScope: { type: string, enum: [all, selected], default: all }
        freeleechMode: { type: string, enum: [honor, bypass], description: "defaults by kind (qui = bypass; *arrs = honor)" }
        priority: { type: integer, default: 25 }
        syncProfileId:
          type: integer
          format: int64
          nullable: true
          description: "optional sync profile reference; must exist, must not be set for qui, and a non-empty category set must overlap the app's content range (else 400)"
      required: [name, kind]
    UpdateConnection:
      type: object
      description: >
        A partial update of the connection's surface fields; omitted fields are
        unchanged. Identity + credential (base URL, api key, harbrr URL) are App-level —
        rotated via the App (PATCH /api/apps/{id}), not here.
      properties:
        name: { type: string }
        syncLevel: { type: string, enum: [full, add_update] }
        indexScope: { type: string, enum: [all, selected] }
        freeleechMode: { type: string, enum: [honor, bypass] }
        priority: { type: integer }
        syncProfileId:
          type: integer
          format: int64
          nullable: true
          description: >
            tri-state: omit to leave unchanged, null to clear, an id to set it (validated
            as on create). Applied verbatim once validated.
...
    ConnectionIndexer:
      type: object
      description: one row of a connection's per-indexer sync ledger
      properties:
        instanceId: { type: integer, format: int64 }
        remoteId: { type: string, description: the id the app assigned the pushed indexer }
        selected: { type: boolean }
        lastPushedAt: { type: string, format: date-time }
        lastPushStatus: { type: string, description: "ok | error" }
        lastPushError: { type: string, description: scrubbed error from the last push }
      required: [instanceId, selected]
    ConnectionStatus:
      allOf:
        - { $ref: "#/components/schemas/AppConnection" }
        - type: object
          properties:
            indexers:
              type: array
              items: { $ref: "#/components/schemas/ConnectionIndexer" }
          required: [indexers]
    SyncResult:
      type: object
      properties:
        slug: { type: string }
        action: { type: string, description: "created | updated | noop | deleted | failed" }
        error: { type: string, description: scrubbed error when action=failed }
      required: [slug, action]
    SyncReport:
      type: object
      properties:
        status: { type: string, description: "ok | partial | error | skipped" }
        results:
          type: array
          items: { $ref: "#/components/schemas/SyncResult" }
      required: [status, results]
    ConnectionSyncResult:
      type: object
      properties:
        connectionId: { type: integer, format: int64 }
        name: { type: string }
        report: { $ref: "#/components/schemas/SyncReport" }
        error: { type: string, description: scrubbed error when this connection failed to sync }
      required: [connectionId, name, report]
    SelectIndexers:
      type: object
      properties:
        instanceIds:
          type: array
          items: { type: integer, format: int64 }
          description: the indexer instance ids to select; all others are cleared
      required: [instanceIds]
...